Retry after a delay if the API rate limit is hit
Instead of scaling, or failing nicely we were just blowing up if the API
limit is hit.  This way it will retry a few times and hopefully work.
Though it does not fix the problem of to many API calls in general and
could make the problem worse with retrying.